Implement plugin dependecies, topological sorting
