+++ packages/plugin/src/PluginUtils.test.tsx
@@ -5,7 +5,11 @@ import { dhTruck, vsPreview } from '@deephaven/icons';
 import {
   type DashboardPlugin,
   type ElementPlugin,
+  type LegacyPlugin,
+  type MultiPlugin,
+  type Plugin,
   type PluginModule,
+  type PluginModuleMap,
   PluginType,
   type ThemePlugin,
   type WidgetPlugin,
@@ -15,6 +19,11 @@ import {
   getIconForPlugin,
   getThemeDataFromPlugins,
   getPluginsElementMap,
+  getPluginModuleValue,
+  processLoadedModule,
+  registerPlugin,
+  sortPluginsByDependency,
+  type PluginManifestPluginInfo,
 } from './PluginUtils';
 
 function TestWidget() {
@@ -208,13 +217,530 @@ describe('getElementPluginMap', () => {
   });
 
   it('should return an empty map if no element plugins are present', () => {
-    const pluginMap = new Map<string, PluginModule>([
+    const pluginMap2 = new Map<string, PluginModule>([
       [widgetPlugin.name, widgetPlugin],
       [dashboardPlugin.name, dashboardPlugin],
     ]);
 
-    const elementMapping = getPluginsElementMap(pluginMap);
+    const elementMapping = getPluginsElementMap(pluginMap2);
 
     expect(elementMapping.size).toBe(0);
+  });
+});
+
+describe('getPluginModuleValue', () => {
+  const legacyPlugins: [type: string, moduleValue: LegacyPlugin][] = [
+    [
+      'dashboard',
+      {
+        DashboardPlugin: () => null,
+      },
+    ],
+    [
+      'auth',
+      {
+        AuthPlugin: {
+          Component: () => null,
+          isAvailable: () => true,
+        },
+      },
+    ],
+    [
+      'table',
+      {
+        TablePlugin: () => null,
+      },
+    ],
+  ];
+
+  const newPlugins: [type: string, moduleValue: Plugin][] = Object.keys(
+    PluginType
+  ).map(type => [type, { name: `${type}`, type: PluginType[type] }]);
+
+  const newPluginsWithNamedExports: [
+    type: string,
+    moduleValue: { default: Plugin; [key: string]: unknown },
+  ][] = Object.keys(PluginType).map(type => [
+    type,
+    {
+      default: { name: `${type}Plugin`, type: PluginType[type] },
+      NamedExport: 'NamedExportValue',
+    },
+  ]);
+
+  const combinedPlugins: [
+    type: string,
+    moduleValue: {
+      default: Plugin;
+    } & LegacyPlugin,
+  ][] = [
+    [
+      'dashboard',
+      {
+        default: {
+          name: 'combinedFormat1',
+          type: PluginType.DASHBOARD_PLUGIN,
+        },
+        DashboardPlugin: () => null,
+      },
+    ],
+    [
+      'auth',
+      {
+        default: {
+          name: 'combinedFormat2',
+          type: PluginType.AUTH_PLUGIN,
+        },
+        AuthPlugin: {
+          Component: () => null,
+          isAvailable: () => true,
+        },
+      },
+    ],
+    [
+      'table',
+      {
+        default: {
+          name: 'combinedFormat3',
+          type: PluginType.TABLE_PLUGIN,
+        },
+        TablePlugin: () => null,
+      },
+    ],
+    [
+      // Should be able to combine different plugin types
+      'multiple',
+      {
+        default: {
+          name: 'widgetPlugin2',
+          type: PluginType.WIDGET_PLUGIN,
+        },
+        DashboardPlugin: () => null,
+      },
+    ],
+  ];
+
+  it.each(legacyPlugins)(
+    'supports legacy %s plugin format',
+    (type, legacyPlugin) => {
+      const moduleValue = getPluginModuleValue(legacyPlugin);
+      expect(moduleValue).toBe(legacyPlugin);
+    }
+  );
+
+  it.each(newPlugins)('supports new %s format', (type, plugin) => {
+    const moduleValue = getPluginModuleValue(plugin);
+    expect(moduleValue).toBe(plugin);
+  });
+
+  it.each(newPluginsWithNamedExports)(
+    'supports new %s format with named exports',
+    (type, plugin) => {
+      const moduleValue = getPluginModuleValue(plugin);
+      expect(moduleValue).toBe(plugin.default);
+    }
+  );
+
+  it.each(combinedPlugins)(
+    'prioritizes new %s plugin if the module contains both legacy and new format',
+    (type, plugin) => {
+      const moduleValue = getPluginModuleValue(plugin);
+      expect(moduleValue).toBe(plugin.default);
+    }
+  );
+
+  it('returns null if the module value is not a plugin', () => {
+    const moduleValue = getPluginModuleValue({} as Plugin);
+    expect(moduleValue).toBeNull();
+  });
+
+  describe('MultiPlugin', () => {
+    const multiPlugin: MultiPlugin = {
+      name: 'test-multi-plugin',
+      type: PluginType.MULTI_PLUGIN,
+      plugins: [
+        { name: 'widget-plugin', type: PluginType.WIDGET_PLUGIN },
+        { name: 'dashboard-plugin2', type: PluginType.DASHBOARD_PLUGIN },
+        { name: 'theme-plugin', type: PluginType.THEME_PLUGIN },
+      ] as Plugin[],
+    };
+
+    it('supports MultiPlugin format', () => {
+      const moduleValue = getPluginModuleValue(multiPlugin);
+      expect(moduleValue).toBe(multiPlugin);
+    });
+
+    it('supports MultiPlugin with default export', () => {
+      const moduleWithDefault = { default: multiPlugin };
+      const moduleValue = getPluginModuleValue(moduleWithDefault);
+      expect(moduleValue).toBe(multiPlugin);
+    });
+
+    it('supports MultiPlugin with named exports', () => {
+      const moduleWithNamedExports = {
+        default: multiPlugin,
+        SomeNamedExport: 'value',
+      };
+      const moduleValue = getPluginModuleValue(moduleWithNamedExports);
+      expect(moduleValue).toBe(multiPlugin);
+    });
+  });
+});
+
+describe('registerPlugin', () => {
+  function makePlugin(name: string): Plugin {
+    return { name, type: PluginType.WIDGET_PLUGIN };
+  }
+
+  it('registers a plugin in the map', () => {
+    const regPluginMap: PluginModuleMap = new Map();
+    const plugin = makePlugin('my-plugin');
+
+    registerPlugin(regPluginMap, 'my-plugin', plugin, '1.0.0');
+
+    expect(regPluginMap.get('my-plugin')).toEqual({
+      ...plugin,
+      version: '1.0.0',
+    });
+  });
+
+  it('skips duplicate and keeps the first registration', () => {
+    const regPluginMap: PluginModuleMap = new Map();
+    const first = makePlugin('my-plugin');
+    const second = makePlugin('my-plugin');
+
+    registerPlugin(regPluginMap, 'my-plugin', first, '1.0.0');
+    registerPlugin(regPluginMap, 'my-plugin', second, '2.0.0');
+
+    expect(regPluginMap.size).toBe(1);
+    expect(regPluginMap.get('my-plugin')).toEqual({
+      ...first,
+      version: '1.0.0',
+    });
+  });
+});
+
+describe('processLoadedModule', () => {
+  function makePlugin(name: string): Plugin {
+    return { name, type: PluginType.WIDGET_PLUGIN };
+  }
+
+  let procPluginMap: PluginModuleMap;
+  let resolveMap: Record<string, unknown>;
+
+  beforeEach(() => {
+    procPluginMap = new Map();
+    resolveMap = {};
+  });
+
+  it('registers a simple plugin', () => {
+    const plugin = makePlugin('my-plugin');
+    processLoadedModule(
+      procPluginMap,
+      resolveMap,
+      plugin,
+      'my-plugin',
+      null,
+      '1.0.0'
+    );
+
+    expect(procPluginMap.size).toBe(1);
+    expect(procPluginMap.get('my-plugin')).toEqual({
+      ...plugin,
+      version: '1.0.0',
+    });
+  });
+
+  it('registers a plugin with default export', () => {
+    const plugin = makePlugin('my-plugin');
+    processLoadedModule(
+      procPluginMap,
+      resolveMap,
+      { default: plugin },
+      'my-plugin'
+    );
+
+    expect(procPluginMap.size).toBe(1);
+    expect(procPluginMap.get('my-plugin')).toEqual({
+      ...plugin,
+      version: undefined,
+    });
+  });
+
+  it('registers in resolve map when packageName is provided', () => {
+    const plugin = makePlugin('my-plugin');
+    const pluginExports = { default: plugin };
+    processLoadedModule(
+      procPluginMap,
+      resolveMap,
+      pluginExports,
+      'my-plugin',
+      '@scope/my-plugin'
+    );
+
+    expect(resolveMap['@scope/my-plugin']).toBe(pluginExports);
+  });
+
+  it('does not register in resolve map when packageName is null', () => {
+    const plugin = makePlugin('my-plugin');
+    processLoadedModule(procPluginMap, resolveMap, plugin, 'my-plugin', null);
+
+    expect(Object.keys(resolveMap)).toHaveLength(0);
+  });
+
+  it('does not register in resolve map when packageName is undefined', () => {
+    const plugin = makePlugin('my-plugin');
+    processLoadedModule(procPluginMap, resolveMap, plugin, 'my-plugin');
+
+    expect(Object.keys(resolveMap)).toHaveLength(0);
+  });
+
+  it('flattens MultiPlugin and registers inner plugins', () => {
+    const multi: MultiPlugin = {
+      name: 'multi',
+      type: PluginType.MULTI_PLUGIN,
+      plugins: [makePlugin('inner-a'), makePlugin('inner-b')] as Plugin[],
+    };
+
+    processLoadedModule(
+      procPluginMap,
+      resolveMap,
+      multi,
+      'multi',
+      null,
+      '1.0.0'
+    );
+
+    expect(procPluginMap.size).toBe(2);
+    expect(procPluginMap.has('inner-a')).toBe(true);
+    expect(procPluginMap.has('inner-b')).toBe(true);
+    expect(procPluginMap.has('multi')).toBe(false);
+  });
+
+  it('skips invalid inner plugins in MultiPlugin', () => {
+    const multi: MultiPlugin = {
+      name: 'multi',
+      type: PluginType.MULTI_PLUGIN,
+      plugins: [
+        makePlugin('valid'),
+        { notAPlugin: true } as unknown as Plugin,
+      ] as Plugin[],
+    };
+
+    processLoadedModule(procPluginMap, resolveMap, multi, 'multi');
+
+    expect(procPluginMap.size).toBe(1);
+    expect(procPluginMap.has('valid')).toBe(true);
+  });
+
+  it('skips inner plugins with empty names in MultiPlugin', () => {
+    const multi: MultiPlugin = {
+      name: 'multi',
+      type: PluginType.MULTI_PLUGIN,
+      plugins: [
+        makePlugin('valid'),
+        { name: '', type: PluginType.WIDGET_PLUGIN } as Plugin,
+        { name: '  ', type: PluginType.WIDGET_PLUGIN } as Plugin,
+      ] as Plugin[],
+    };
+
+    processLoadedModule(procPluginMap, resolveMap, multi, 'multi');
+
+    expect(procPluginMap.size).toBe(1);
+    expect(procPluginMap.has('valid')).toBe(true);
+  });
+
+  it('skips duplicate inner plugins in MultiPlugin', () => {
+    const multi: MultiPlugin = {
+      name: 'multi',
+      type: PluginType.MULTI_PLUGIN,
+      plugins: [makePlugin('dupe'), makePlugin('dupe')] as Plugin[],
+    };
+
+    processLoadedModule(procPluginMap, resolveMap, multi, 'multi');
+
+    expect(procPluginMap.size).toBe(1);
+  });
+
+  it('does not register when module value is null', () => {
+    processLoadedModule(procPluginMap, resolveMap, {} as Plugin, 'bad-plugin');
+
+    expect(procPluginMap.size).toBe(0);
+  });
+
+  it('handles legacy plugin format', () => {
+    const legacy: LegacyPlugin = { TablePlugin: () => null };
+    processLoadedModule(
+      procPluginMap,
+      resolveMap,
+      legacy,
+      'legacy-plugin',
+      null,
+      '1.0.0'
+    );
+
+    expect(procPluginMap.size).toBe(1);
+    expect(procPluginMap.get('legacy-plugin')).toEqual({
+      ...legacy,
+      version: '1.0.0',
+    });
+  });
+});
+
+describe('sortPluginsByDependency', () => {
+  function makeManifestPlugin(
+    name: string,
+    opts?: {
+      package?: string;
+      dependencies?: string[];
+    }
+  ): PluginManifestPluginInfo {
+    return {
+      name,
+      main: 'index.js',
+      version: '1.0.0',
+      package: opts?.package,
+      dependencies: opts?.dependencies,
+    };
+  }
+
+  it('returns plugins in original order when no dependencies', () => {
+    const plugins = [
+      makeManifestPlugin('a'),
+      makeManifestPlugin('b'),
+      makeManifestPlugin('c'),
+    ];
+
+    const sorted = sortPluginsByDependency(plugins);
+
+    expect(sorted.map(p => p.name)).toEqual(['a', 'b', 'c']);
+  });
+
+  it('reorders so dependency loads before consumer', () => {
+    const plugins = [
+      makeManifestPlugin('consumer', {
+        dependencies: ['@scope/dep'],
+      }),
+      makeManifestPlugin('dep', {
+        package: '@scope/dep',
+      }),
+    ];
+
+    const sorted = sortPluginsByDependency(plugins);
+
+    expect(sorted.map(p => p.name)).toEqual(['dep', 'consumer']);
+  });
+
+  it('handles a chain of dependencies: a → b → c', () => {
+    const plugins = [
+      makeManifestPlugin('c', {
+        package: '@scope/c',
+        dependencies: ['@scope/b'],
+      }),
+      makeManifestPlugin('b', {
+        package: '@scope/b',
+        dependencies: ['@scope/a'],
+      }),
+      makeManifestPlugin('a', {
+        package: '@scope/a',
+      }),
+    ];
+
+    const sorted = sortPluginsByDependency(plugins);
+
+    const names = sorted.map(p => p.name);
+    expect(names.indexOf('a')).toBeLessThan(names.indexOf('b'));
+    expect(names.indexOf('b')).toBeLessThan(names.indexOf('c'));
+  });
+
+  it('preserves original order among independent plugins', () => {
+    const plugins = [
+      makeManifestPlugin('x'),
+      makeManifestPlugin('dep', { package: '@scope/dep' }),
+      makeManifestPlugin('y'),
+      makeManifestPlugin('consumer', { dependencies: ['@scope/dep'] }),
+      makeManifestPlugin('z'),
+    ];
+
+    const sorted = sortPluginsByDependency(plugins);
+
+    const names = sorted.map(p => p.name);
+    // dep must come before consumer
+    expect(names.indexOf('dep')).toBeLessThan(names.indexOf('consumer'));
+    // independent plugins keep their relative order
+    expect(names.indexOf('x')).toBeLessThan(names.indexOf('y'));
+    expect(names.indexOf('y')).toBeLessThan(names.indexOf('z'));
+  });
+
+  it('throws on circular dependencies', () => {
+    const plugins = [
+      makeManifestPlugin('a', {
+        package: '@scope/a',
+        dependencies: ['@scope/b'],
+      }),
+      makeManifestPlugin('b', {
+        package: '@scope/b',
+        dependencies: ['@scope/a'],
+      }),
+    ];
+
+    expect(() => sortPluginsByDependency(plugins)).toThrow(
+      /Circular plugin dependency/
+    );
+  });
+
+  it('warns and ignores dependencies not in the manifest', () => {
+    const plugins = [
+      makeManifestPlugin('consumer', {
+        dependencies: ['@scope/nonexistent'],
+      }),
+    ];
+
+    const sorted = sortPluginsByDependency(plugins);
+
+    expect(sorted.map(p => p.name)).toEqual(['consumer']);
+  });
+
+  it('handles multiple dependencies', () => {
+    const plugins = [
+      makeManifestPlugin('consumer', {
+        dependencies: ['@scope/dep-a', '@scope/dep-b'],
+      }),
+      makeManifestPlugin('dep-a', { package: '@scope/dep-a' }),
+      makeManifestPlugin('dep-b', { package: '@scope/dep-b' }),
+    ];
+
+    const sorted = sortPluginsByDependency(plugins);
+
+    const names = sorted.map(p => p.name);
+    expect(names.indexOf('dep-a')).toBeLessThan(names.indexOf('consumer'));
+    expect(names.indexOf('dep-b')).toBeLessThan(names.indexOf('consumer'));
+  });
+
+  it('does not mutate the input array', () => {
+    const plugins = [
+      makeManifestPlugin('consumer', { dependencies: ['@scope/dep'] }),
+      makeManifestPlugin('dep', { package: '@scope/dep' }),
+    ];
+    const original = [...plugins];
+
+    sortPluginsByDependency(plugins);
+
+    expect(plugins).toEqual(original);
+  });
+
+  it('handles empty plugin list', () => {
+    expect(sortPluginsByDependency([])).toEqual([]);
+  });
+
+  it('handles plugins with empty dependencies array', () => {
+    const plugins = [
+      makeManifestPlugin('a', { dependencies: [] }),
+      makeManifestPlugin('b'),
+    ];
+
+    const sorted = sortPluginsByDependency(plugins);
+
+    expect(sorted.map(p => p.name)).toEqual(['a', 'b']);
   });
 });
